requirements.txt
mws==0.8.6
python-dateutil
requests
lxml
